Changes the timeout of the websocket connection
This changes the timeout of the websocket connection to 10 seconds if there is no token stored for a connection.
I have also added a print statement to print the json output. I added a _read_response() to the end of the contol method and removed the creating of an exception from the _read_response method if the data is not a connect data. this way it will also print out data for key presses