adds upnp functions. reworks the websocketssl
everything is all automatic now. no need for config files. or having a
static IP on the TV.

its as easy as.

from samsungctl import remote

for tv in remote.discover(timeout=5):
    print(tv.volume)

There are bound to be bugs. sp please report them so I can get it fixed.

I have an older samsung TV (2011) I do not know what features are
available on the newer ones. I only coded based on what is on mine.
if you run the example.py file it is going to print out a whole lot of
information. It is also going change settings on your TV to verify that
things work. call it a small demo if you will.

If you can copy and past the output I will be able to see what is
available on the newer TV's

Remember run the example.py file.